Added C++ message_decode example using FusionEngineFramer.
Renamed the previous example to raw_message_decode.
