+++ examples/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(common)
 add_subdirectory(generate_data)
 add_subdirectory(lband_decode)
 add_subdirectory(message_decode)
+add_subdirectory(raw_message_decode)
 add_subdirectory(request_version)
 add_subdirectory(tcp_client)
 add_subdirectory(udp_client)
